requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.22.0
azure-storage-blob==12.25.1
bitarray==3.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
cryptography==44.0.3
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.2.2
frozenlist==1.6.0
hexbytes==1.3.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
packaging==25.0
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
tomli==2.2.1
toolz==1.0.0
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
web3==7.11.1
websockets==15.0.1
yarl==1.20.0
